// src/lib/types.ts
export interface JikanRelation {
  relation: string;
  entry: {
    mal_id: number;
    type: string;
    name: string;
    url: string;
  }[];
}

export interface Anime {
  mal_id: number;
  id: number; // AniList ID (or MAL ID as a stand-in)
  url: string;
  images: {
    jpg: {
      image_url: string;
      small_image_url: string;
      large_image_url: string;
    };
    webp: {
      image_url: string;
      small_image_url: string;
      large_image_url: string;
    };
  };
  trailer: {
    youtube_id: string | null;
    url: string | null;
    embed_url: string | null;
  };
  title: string;
  title_english: string | null;
  title_japanese: string;
  synopsis: string;
  type: string; // ANIME or MANGA
  format?: string;
  source: string;
  episodes: number | null; // For Anime
  nextAiringEpisode?: {
    airingAt: number;
    timeUntilAiring: number;
    episode: number;
  } | null;
  chapters: number | null; // For Manga
  volumes: number | null; // For Manga
  status: string;
  airing: boolean;
  score: number | null;
  scored_by: number | null;
  rank: number | null;
  popularity: number | null;
  year: number | null;
  isAdult?: boolean;
  startDate?: {
    year: number | null;
    month: number | null;
    day: number | null;
  };
  broadcast: {
    day: string | null;
    time: string | null;
    timezone: string | null;
    string: string | null;
  };
  rating: string | null;
  genres: JikanGenre[];
  studios: { mal_id: number; type: string; name: string; url: string }[];
  authors?: { mal_id: number; name: string; url: string, images: { jpg: { image_url: string } } }[];
  relations?: JikanRelation[];
  relationType?: string; // Added for watch order context
  update?: {
    newEpisodes: number;
  }
  latestUpdate?: {
    status: string;
    progress: string;
    timestamp: number;
  }
}
// ...
export interface JikanVoiceActor {
    person: {
        mal_id: number;
        url?: string;
        images: { jpg: { image_url: string } };
        name: string;
    };
    language: string;
}

export interface JikanCharacterDetail {
    mal_id: number;
    name: string;
    name_kanji: string;
    nicknames: string[];
    favorites: number;
    about: string;
    images: {
        jpg: { image_url: string };
        webp: { image_url: string };
    };
    anime: {
        role: string;
        anime: Anime;
    }[];
    voices: JikanVoiceActor[];
}
// ...
export interface JikanGenre {
    mal_id: number;
    name: string;
    url?: string;
    count?: number;
    type: 'genre' | 'tag';
}

export interface JikanAPIGenresResponse {
    data: JikanGenre[];
}

// Pictures
export interface JikanPicture {
    jpg: { image_url: string };
    webp?: { image_url: string };
}